Bug fixes and changes to `full_types_manager` (for EOSIO/eos#354):   * Fixed bug where a missing static keyword led to undefined behavior     which happen to result in all sorted fields being treated as if     they were all specified in the ascending order.   * Now `full_types_manager` has field names and sort order information.   * This enables more informative printing of types, which was in fact     implemented along with making the printed struct formatting look nicer.   * While printing the structure of types is not super important, the above     changes provide the framework necessary for conversion to/from JSON     using just: the raw data, the `type_id` associated with that raw data,     and the information available within the `full_types_manager` class.

+++ libraries/types/include/eos/types/types_manager_common.hpp
@@ -345,7 +345,7 @@ namespace eos { namespace types {
       const vector<uint32_t>&       types;
       const vector<field_metadata>& members;
 
-      tuple<uint32_t, uint32_t, uint32_t> get_members_common(type_id::index_t struct_index)const;
+      tuple<uint32_t, uint16_t, uint16_t> get_members_common(type_id::index_t struct_index)const;
       type_id::size_align                 get_size_align(type_id tid, uint32_t* cache_ptr)const;
 
    };
